# frontend/packages/ui/src/docs/openapi-authenticated.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/openapi-docs/{id}":{"put":{"tags":["open-api-docs-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRawJsonRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/openapi-docs":{"get":{"tags":["open-api-docs-controller"],"operationId":"listAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OpenApiDocSummary"}}}}}}},"post":{"tags":["open-api-docs-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/api-keys":{"get":{"tags":["api-key-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApiKeyDTO"}}}}}}},"post":{"tags":["api-key-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApiKeyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiKeyDTO"}}}}}}},"/api/api-keys/{id}/revoke":{"post":{"tags":["api-key-controller"],"operationId":"revoke","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/openapi-docs/{id}/summary":{"get":{"tags":["open-api-docs-controller"],"operationId":"getSummary","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OpenApiSummaryDTO"}}}}}}},"/api/openapi-docs/{id}/raw":{"get":{"tags":["open-api-docs-controller"],"operationId":"getRawJson","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/openapi-docs/{id}/endpoints":{"get":{"tags":["open-api-docs-controller"],"operationId":"getEndpoints","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EndpointDTO"}}}}}}},"/api/openapi-docs/cicd-sample":{"get":{"tags":["open-api-docs-controller"],"operationId":"cicdSample","responses":{"200":{"description":"OK","content":{"text/markdown":{"schema":{"type":"string"}}}}}}},"/api/auth/user":{"get":{"tags":["auth-controller"],"operationId":"user","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/auth/status":{"get":{"tags":["auth-controller"],"operationId":"authStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/api-keys/{id}":{"delete":{"tags":["api-key-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UpdateRawJsonRequest":{"type":"object","properties":{"rawJson":{"type":"string"}}},"CreateRequest":{"type":"object","properties":{"serviceName":{"type":"string"},"rawJson":{"type":"string"}}},"CreateApiKeyRequest":{"type":"object","properties":{"name":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"}}},"ApiKeyDTO":{"type":"object","properties":{"id":{"type":"string"},"ownerId":{"$ref":"#/components/schemas/UserId"},"name":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"expiresAt":{"type":"string","format":"date-time"},"lastUsedAt":{"type":"string","format":"date-time"},"revoked":{"type":"boolean"},"valid":{"type":"boolean"}}},"UserId":{},"OpenApiDocSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ownerId":{"type":"string"},"serviceName":{"type":"string"},"version":{"type":"integer","format":"int64"},"hash":{"type":"string"}}},"OpenApiSummaryDTO":{"type":"object","properties":{"title":{"type":"string"},"version":{"type":"string"},"description":{"type":"string"},"totalEndpoints":{"type":"integer","format":"int32"},"tags":{"type":"array","items":{"type":"string"}}}},"EndpointDTO":{"type":"object","properties":{"summary":{"$ref":"#/components/schemas/EndpointSummaryDTO"},"endpoints":{"type":"array","items":{"$ref":"#/components/schemas/EndpointItemDTO"}}}},"EndpointItemDTO":{"type":"object","properties":{"path":{"type":"string"},"method":{"type":"string","enum":["GET","POST","PUT","PATCH","DELETE","HEAD","OPTIONS"]},"operationId":{"type":"string"},"summary":{"type":"string"},"description":{"type":"string"},"parameters":{"type":"array","items":{"$ref":"#/components/schemas/EndpointParameter"}},"requestBody":{"$ref":"#/components/schemas/EndpointRequestBodyDTO"},"responses":{"type":"array","items":{"$ref":"#/components/schemas/EndpointResponseBodyDTO"}},"tags":{"type":"array","items":{"type":"string"}}}},"EndpointParameter":{"type":"object","properties":{"name":{"type":"string"},"location":{"type":"string","enum":["PATH","QUERY","HEADER","COOKIE"]},"type":{"type":"string"},"required":{"type":"boolean"}}},"EndpointRequestBodyDTO":{"type":"object","properties":{"mediaType":{"type":"string"},"schemaJson":{"type":"string"},"required":{"type":"boolean"}}},"EndpointResponseBodyDTO":{"type":"object","properties":{"statusCode":{"type":"integer","format":"int32"},"mediaType":{"type":"string"},"schemaJson":{"type":"string"},"description":{"type":"string"}}},"EndpointSummaryDTO":{"type":"object","properties":{"docId":{"$ref":"#/components/schemas/OpenApiDocId"},"totalEndpoints":{"type":"integer","format":"int32"}}},"OpenApiDocId":{"type":"object","properties":{"value":{"type":"integer","format":"int64"}}},"UserDTO":{"type":"object","properties":{"id":{"$ref":"#/components/schemas/UserId"},"name":{"type":"string"},"email":{"type":"string"},"picture":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"createdAt":{"type":"string","format":"date-time"},"lastLoginAt":{"type":"string","format":"date-time"}}}}}}