# src/main/java/io/joshuasalcedo/devcom/openapidocs/api/dto/EndpointParameter.java
package io.joshuasalcedo.devcom.openapidocs.api.dto;

import jakarta.persistence.Embeddable;
import jakarta.persistence.EnumType;
import jakarta.persistence.Enumerated;

@Embeddable
public record EndpointParameter(
        String name,
        @Enumerated(EnumType.STRING)
        EndpointParameterLocation location,
        String type,
        boolean required
) {
}